requirements.txt
python-binance==1.0.19
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
ta==0.10.2
websocket-client==1.7.0